AI Society/project-manager-realm/intelligence/service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# LLM & AI
google-genai==0.3.0
anthropic==0.7.0
openai==1.3.0

# Vector Database
chromadb==0.4.21
pinecone-client==3.0.2

# Event Streaming
aiokafka==0.10.0
confluent-kafka==2.3.0

# Observability & Monitoring
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==0.42b0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
testcontainers==3.7.1

# Development
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
types-python-dateutil==2.8.19.14

# Utilities
pyyaml==6.0.1
jinja2==3.1.2
tiktoken==0.5.2
